api/requirements.txt
fastapi==0.122.0
pydantic==2.12.5
pyyaml==6.0.3
redis==7.1.0
uvicorn==0.38.0
websockets==15.0.1
